[AutoDiff] LLVM adstack lazy-claim: split into stage-grouped subdir (bound_eval / metadata_publish / heap_grow)
